@@ -1,111 +1,93 @@
 {
-  "name": "my-project",
-  "version": "0.1.0",
+  "name": "mall4cloud",
   "private": true,
+  "version": "0.0.0",
   "scripts": {
-    "serve": "npm run dev:h5",
-    "build": "npm run build:h5",
-    "build:app-plus": "cross-env NODE_ENV=production UNI_PLATFORM=app-plus vue-cli-service uni-build",
-    "build:custom": "cross-env NODE_ENV=production uniapp-cli custom",
-    "build:h5": "cross-env NODE_ENV=production UNI_PLATFORM=h5 vue-cli-service uni-build",
-    "build:mp-360": "cross-env NODE_ENV=production UNI_PLATFORM=mp-360 vue-cli-service uni-build",
-    "build:mp-alipay": "cross-env NODE_ENV=production UNI_PLATFORM=mp-alipay vue-cli-service uni-build",
-    "build:mp-baidu": "cross-env NODE_ENV=production UNI_PLATFORM=mp-baidu vue-cli-service uni-build",
-    "build:mp-kuaishou": "cross-env NODE_ENV=production UNI_PLATFORM=mp-kuaishou vue-cli-service uni-build",
-    "build:mp-qq": "cross-env NODE_ENV=production UNI_PLATFORM=mp-qq vue-cli-service uni-build",
-    "build:mp-toutiao": "cross-env NODE_ENV=production UNI_PLATFORM=mp-toutiao vue-cli-service uni-build",
-    "build:mp-weixin": "cross-env NODE_ENV=production UNI_PLATFORM=mp-weixin vue-cli-service uni-build",
-    "build:quickapp-native": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-native vue-cli-service uni-build",
-    "build:quickapp-webview": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-webview vue-cli-service uni-build",
-    "build:quickapp-webview-huawei": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-webview-huawei vue-cli-service uni-build",
-    "build:quickapp-webview-union": "cross-env NODE_ENV=production UNI_PLATFORM=quickapp-webview-union vue-cli-service uni-build",
-    "dev:app-plus": "cross-env NODE_ENV=development UNI_PLATFORM=app-plus vue-cli-service uni-build --watch",
-    "dev:custom": "cross-env NODE_ENV=development uniapp-cli custom",
-    "dev:h5": "cross-env NODE_ENV=development UNI_PLATFORM=h5 vue-cli-service uni-serve",
-    "dev:mp-360": "cross-env NODE_ENV=development UNI_PLATFORM=mp-360 vue-cli-service uni-build --watch",
-    "dev:mp-alipay": "cross-env NODE_ENV=development UNI_PLATFORM=mp-alipay vue-cli-service uni-build --watch",
-    "dev:mp-baidu": "cross-env NODE_ENV=development UNI_PLATFORM=mp-baidu vue-cli-service uni-build --watch",
-    "dev:mp-kuaishou": "cross-env NODE_ENV=development UNI_PLATFORM=mp-kuaishou vue-cli-service uni-build --watch",
-    "dev:mp-qq": "cross-env NODE_ENV=development UNI_PLATFORM=mp-qq vue-cli-service uni-build --watch",
-    "dev:mp-toutiao": "cross-env NODE_ENV=development UNI_PLATFORM=mp-toutiao vue-cli-service uni-build --watch",
-    "dev:mp-weixin": "cross-env NODE_ENV=development UNI_PLATFORM=mp-weixin vue-cli-service uni-build --watch",
-    "dev:quickapp-native": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-native vue-cli-service uni-build --watch",
-    "dev:quickapp-webview": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-webview vue-cli-service uni-build --watch",
-    "dev:quickapp-webview-huawei": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-webview-huawei vue-cli-service uni-build --watch",
-    "dev:quickapp-webview-union": "cross-env NODE_ENV=development UNI_PLATFORM=quickapp-webview-union vue-cli-service uni-build --watch",
-    "info": "node node_modules/@dcloudio/vue-cli-plugin-uni/commands/info.js",
-    "serve:quickapp-native": "node node_modules/@dcloudio/uni-quickapp-native/bin/serve.js",
-    "test:android": "cross-env UNI_PLATFORM=app-plus UNI_OS_NAME=android jest -i",
-    "test:h5": "cross-env UNI_PLATFORM=h5 jest -i",
-    "test:ios": "cross-env UNI_PLATFORM=app-plus UNI_OS_NAME=ios jest -i",
-    "test:mp-baidu": "cross-env UNI_PLATFORM=mp-baidu jest -i",
-    "test:mp-weixin": "cross-env UNI_PLATFORM=mp-weixin jest -i"
+    "preinstall": "npx only-allow pnpm",
+    "dev:h5": "uni",
+    "dev:mp-weixin": "uni -p mp-weixin",
+    "dev:mp-weixin-test": "uni -p mp-weixin --mode testing",
+    "build:app-android": "uni build -p app-android",
+    "build:app-android-test": "uni build -p app-android --mode testing --outDir=./dist/test/app",
+    "build:app-ios": "uni build -p app-ios",
+    "build:h5": "uni build",
+    "build:mp-weixin": "uni build -p mp-weixin",
+    "build:mp-weixin-test": "uni build -p mp-weixin --mode testing --outDir=./dist/test/mp-weixin",
+    "lint": "eslint --ext .js,.vue src",
+    "lint:fix": "eslint --fix --ext .js --ext .jsx --ext .vue src",
+    "prepare": "husky install",
+    "lint:staged": "lint-staged"
   },
   "dependencies": {
-    "@dcloudio/uni-app-plus": "^2.0.1-33920220314002",
-    "@dcloudio/uni-h5": "^2.0.1-33920220314002",
-    "@dcloudio/uni-helper-json": "^1.0.12",
-    "@dcloudio/uni-i18n": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-360": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-alipay": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-baidu": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-jd": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-kuaishou": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-lark": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-qq": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-toutiao": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-vue": "^2.0.1-33920220314002",
-    "@dcloudio/uni-mp-weixin": "^2.0.1-33920220314002",
-    "@dcloudio/uni-quickapp-native": "^2.0.1-33920220314002",
-    "@dcloudio/uni-quickapp-webview": "^2.0.1-33920220314002",
-    "@dcloudio/uni-stat": "^2.0.1-33920220314002",
-    "@vue/shared": "^3.2.31",
-    "axios": "^0.21.1",
-    "core-js": "^3.6.5",
-    "eslint": "^7.18.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-html": "^6.1.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^5.0.0",
-    "eslint-plugin-vue": "^7.5.0",
-    "flyio": "^0.6.14",
-    "minio": "^7.0.18",
-    "regenerator-runtime": "^0.12.1",
-    "uni-read-pages": "^1.0.5",
-    "uni-simple-router": "^2.0.5",
-    "vue": "^2.6.11",
-    "vue-i18n": "^8.21.0",
-    "vuex": "^3.6.2"
+    "@babel/eslint-parser": "7.23.3",
+    "@dcloudio/uni-app": "3.0.0-3080720230703001",
+    "@dcloudio/uni-app-plus": "3.0.0-3080720230703001",
+    "@dcloudio/uni-components": "3.0.0-3080720230703001",
+    "@dcloudio/uni-h5": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-alipay": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-baidu": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-jd": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-kuaishou": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-lark": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-qq": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-toutiao": "3.0.0-3080720230703001",
+    "@dcloudio/uni-mp-weixin": "3.0.0-3080720230703001",
+    "@dcloudio/uni-quickapp-webview": "3.0.0-3080720230703001",
+    "@uni-ui/code-ui": "1.5.3",
+    "big.js": "6.2.1",
+    "crypto-js": "4.2.0",
+    "flv.js": "^1.6.2",
+    "js-base64": "3.7.5",
+    "moment": "2.29.4",
+    "uni-crazy-router": "1.1.3",
+    "uni-vite-plugin-h5-prod-effect": "1.0.1",
+    "video.js": "8.6.1",
+    "videojs-flvjs-es6": "^1.0.1",
+    "vue": "3.3.8",
+    "vue-i18n": "9.6.5"
   },
   "devDependencies": {
-    "@dcloudio/types": "*",
-    "@dcloudio/uni-automator": "^2.0.1-33920220314002",
-    "@dcloudio/uni-cli-i18n": "^2.0.1-33920220314002",
-    "@dcloudio/uni-cli-shared": "^2.0.1-33920220314002",
-    "@dcloudio/uni-migration": "^2.0.1-33920220314002",
-    "@dcloudio/uni-template-compiler": "^2.0.1-33920220314002",
-    "@dcloudio/vue-cli-plugin-hbuilderx": "^2.0.1-33920220314002",
-    "@dcloudio/vue-cli-plugin-uni": "^2.0.1-33920220314002",
-    "@dcloudio/vue-cli-plugin-uni-optimize": "^2.0.1-33920220314002",
-    "@dcloudio/webpack-uni-mp-loader": "^2.0.1-33920220314002",
-    "@dcloudio/webpack-uni-pages-loader": "^2.0.1-33920220314002",
-    "@vue/cli-plugin-babel": "^4.5.0",
-    "@vue/cli-service": "^4.5.0",
-    "babel-plugin-import": "^1.13.3",
-    "cross-env": "^7.0.3",
-    "jest": "^25.5.4",
-    "mini-types": "*",
-    "miniprogram-api-typings": "*",
-    "postcss-comment": "^2.0.0",
-    "vue-template-compiler": "^2.6.11"
+    "@dcloudio/types": "3.4.3",
+    "@dcloudio/uni-automator": "3.0.0-3080720230703001",
+    "@dcloudio/uni-cli-shared": "3.0.0-3080720230703001",
+    "@dcloudio/uni-stacktracey": "3.0.0-3080720230703001",
+    "@dcloudio/vite-plugin-uni": "3.0.0-3080720230703001",
+    "@vitejs/plugin-legacy": "4.1.1",
+    "eslint": "8.53.0",
+    "eslint-config-standard": "17.1.0",
+    "eslint-plugin-import": "2.29.0",
+    "eslint-plugin-n": "16.3.1",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-vue": "9.18.1",
+    "eslint-plugin-vue-scoped-css": "2.5.1",
+    "husky": "8.0.3",
+    "lint-staged": "15.1.0",
+    "sass": "1.69.5",
+    "terser": "5.24.0",
+    "unplugin-auto-import": "0.16.7",
+    "unplugin-vue-components": "0.25.2",
+    "vite": "4.5.0",
+    "vue-eslint-parser": "9.3.2"
   },
-  "browserslist": [
-    "Android >= 4",
-    "ios >= 8"
-  ],
-  "uni-app": {
-    "scripts": {}
+  "lint-staged": {
+    "*.{js,vue}": [
+      "eslint --fix"
+    ]
+  },
+  "engines": {
+    "node": ">=18.12.0",
+    "pnpm": ">=7"
+  },
+  "pnpm": {
+    "peerDependencyRules": {
+      "ignoreMissing": [
+        "@babel/core*",
+        "jest-environment-node*",
+        "postcss*",
+        "regenerator-runtime*",
+        "url*",
+        "jest*"
+      ]
+    }
   }
 }
